+++ example/lib/model.dart
@@ -1,0 +1,43 @@
+// example viewmodel for the form
+class PonyModel {
+  String name = 'Twilight Sparkle';
+  String type = 'Unicorn';
+  int age = 7;
+  String coatColor = 'D19FE4';
+  String maneColor = '273873';
+  bool hasSpots = false;
+  String spotColor = 'FF5198';
+  String description =
+      'An intelligent and dutiful scholar with an avid love of learning and skill in unicorn magic such as levitation, teleportation, and the creation of force fields.';
+  List<String> hobbies = <String>[
+    'flying',
+    'singing',
+    'exploring',
+    'hiding',
+    'coloring'
+  ];
+  double height = 3.5;
+  int weight = 45;
+  DateTime showDateTime = DateTime(2010, 10, 10, 20, 30);
+  double ticketPrice = 65.99;
+  int boxOfficePhone = 8005551212;
+  String email = 'me@nowhere.org';
+  String password = 'secret1';
+}
+
+const List<String> allHobbies = <String>[
+  'running',
+  'flying',
+  'coloring',
+  'jumping',
+  'eating',
+  'hiding',
+  'exploring',
+  'singing',
+  'dancing',
+  'acting',
+  'cleaning',
+  'shopping',
+  'sewing',
+  'cooking',
+];