+++ lib/card_settings.dart
@@ -27,6 +27,9 @@ export 'widgets/picker_fields/card_settings_list_picker.dart';
 export 'widgets/picker_fields/card_settings_number_picker.dart';
 export 'widgets/picker_fields/card_settings_time_picker.dart';
 
+// selection fields
+export 'widgets/selection_fields/card_settings_multiselect.dart';
+
 // text fields
 export 'widgets/text_fields/card_settings_email.dart';
 export 'widgets/text_fields/card_settings_paragraph.dart';