+++ lib/widgets/picker_fields/card_settings_color_picker.dart
@@ -154,7 +154,7 @@ class _CardSettingsColorPickerState extends FormFieldState<Color> {
     ).then((value) {
       if (value != null) {
         didChange(value);
-        widget.onChanged(value);
+        if (widget.onChanged != null) widget.onChanged(value);
       }
     });
   }