+++ lib/widgets/picker_fields/card_settings_date_picker.dart
@@ -71,7 +71,7 @@ class _CardSettingsDatePickerState extends FormFieldState<DateTime> {
     ).then((value) {
       if (value != null) {
         didChange(value);
-        widget.onChanged(value);
+        if (widget.onChanged != null) widget.onChanged(value);
       }
     });
   }